ai_service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pymongo==4.6.2
motor==3.3.2
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1.post1
networkx==3.2.1
python-dotenv==1.0.1
pydantic==2.6.3
